+++ packages/components/src/Chip/Chip.css.d.ts
@@ -9,6 +9,8 @@ declare const styles: {
   readonly "invalid": string;
   readonly "subtle": string;
   readonly "disabled": string;
+  readonly "chipContent": string;
+  readonly "truncateGradient": string;
 };
 export = styles;